implement: デフォルトのIndexページを、「ダイアログ一覧」になるように、既存のものをすべて削除。
