--- src/components/VuetifyLogo.vue
@@ -1,22 +1,0 @@
-<template>
-  <img
-    class="VuetifyLogo"
-    alt="Vuetify Logo"
-    src="@/static/vuetify-logo.svg"
-  >
-</template>
-
-<style>
-.VuetifyLogo {
-  height: 180px;
-  width: 180px;
-  transform: rotateY(560deg);
-  animation: turn 3.5s ease-out forwards 1s;
-}
-
-@keyframes turn {
-  100% {
-    transform: rotateY(0deg);
-  }
-}
-</style>
